[bazel] add hlcpp visibility support to generated sdk
This CL adds a mechanism which lets us inject visibility into the hlcpp
bindings to restrict their usage. The mechanism at play here lets us
decide visibility on a per SDK basis which is important to make it so we
don't break our OOT users while still restricting the in-tree users.

The CL restricts usage to the fuchsia_sdk and internal_sdk since there
are libraries that are still using hlcpp but it prevents usage in other
locations in-tree.

Bug: 406789405
Change-Id: If427d43585e92019f69e51721805c2f40410f1d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238428
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>