[fdf][rust] Introduce `OnDispatcher` trait
This trait allows for a uniform abraction of a dispatcher that may or
may not be available when it comes time to submit a task to it. This
allows for using a 'Weak<Dispatcher>' directly with channel reads and
task spawning, but also will allow for using the "currently active
dispatcher".

Change-Id: If467bff6e12b6b4748a4e17b3277ffa22329d541
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1223306
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>