[vreg][fidl] Make Voltage interface in vreg FIDL return zx.status
Previously the fuchsia.hardware.vreg interface was only used to set the
voltage of a voltage regulator. As of fxrev.dev/1236544 the interface
can now also be used to enable/disable a regulator.

Some regulators may be fixed-voltage regulators that don't support
programming the voltage but may support the new Enable/Disable interface
only. Those regulators need a way to indicate that they do not support
the SetVoltageStep, GetVoltageStep, and GetRegulatorParams interfaces.

This CL allows those calls to also return a zx.Status so that they can
indicate that the method is not implemented.

Bug: b/406156231
Change-Id: Ib7b38e51b8069de463ebd7500bf7ca6082489734
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237804
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>