[audio][test] Fix virtual-audio-legacy on riscv
Eliminate the usage of a FP method (use static_cast<int>
instead of 'truncf') so as to minimize virtual-audio-legacy
dependence on the C-runtime, for HW targets where this may
be restricted, such as riscv64.

Related: enable 'audio_driver_tests', as well as the CLI
'virtual_audio' utility, even if 'virtual-audio-legacy' is
absent or failing.

Bug: 370741565
Fixed: 406265676

Change-Id: If6345a28af602d0263ea7e3853d54e176193e743
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239046
Reviewed-by: Nick Eaton <nickeaton@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>