[fdf][rust] Improve public dispatcher test helpers
Changes the test helper functions exposed by the fdf env crate to not
rely on using a `DispatcherRef<'static>` that isn't really 'static by
taking advantage of the existence of `CurrentDispatcher`. Also reduces
the complexity of using them by spawning the callback in the dispatcher
and waiting for the callback to finish.

Also, since many tests will likely be async, adds methods to spawn a
future on the dispatcher instead.

Along the way I cleaned up some docs and retracted some things that were
made pub that probably don't need to be.

Change-Id: I80f8dafc80f6ff41ec4be04930d11a4bd9cda687
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238844
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>