[sestarnix] Check sys_rawio for FIBMAP ioctl
This is the last substantial permission-related part of getting the
capable_sys test of the SELinux Test Suite passing. Since it's pretty
obscure (we're not yet aware of any FIBMAP callers other than the
SELinux Test Suite) and we'll need to fake a successful return of the
ioctl we introduce the selinux_test_suite kernel feature flag to allow
Starnix to take action based on whether or not it is running the
SELinux Test Suite.

Bug: 364569559
Bug: 392615438
Bug: 398663315
Bug: 404795644
Change-Id: Idcb31698ebbbcc3a603451b297bf85b9e77ef441
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222964
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Nathaniel Manista <nathaniel@google.com>