[netstack3] Send ICMP Port Unreachable from UDP
Originally, Port Unreachable messages were not sent by the netstack. The
rationale was to avoid contributing to a reflection DOS attack, in which
the attack performs a UDP port scan on the netstack, causing us to send
a large volume of ICMP Port Unreachable errors to the victim.

This attack vector is not a substantial concern, because the netstack's
ICMP implementation enforces a rate limit on the amount of ICMP errors
that can be sent. This is in line with the behavior of other platforms.

Bug: 406666070
Change-Id: I41509e8821e8e3e4e1dff4afff4abba5963535ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239544
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>