[netstack3] Update MSS on PMTU update
Respond to PMTU updates (delivered as ICMP errors to TCP sockets) by
adjusting the sender MSS (and congestion window) accordingly, as well as
eagerly retransmitting up to a single MSS starting at the segment that
was reported to be too big.

Eagerly retransmitting the segment has the advantage of allowing PMTUD
to continue right away, which may prove useful if there are other MTU
adjustments that need to be discovered along the path to the remote.
Note, however, that we *only* retransmit the one segment that was marked
as too big in order to avoid causing a packet storm in this case. Any
other data that must be retransmitted is handled by the retransmission
timer.

Fixed: 404628798
Test: netstack3-core-tcp-test
Test: netstack-socket-integration-test
Change-Id: I0100a0f55d2ded7333eba6e13541a4bcbc630f03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233389
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>