[bazel] require target name in visibility_templates
When a user uses wild card expansion in the visibility_templates
attribute on the fuchsia_sdk_repository rule we need to ensure that they
have specified a target. If they add the '*' at the end of the label it
will try to use '*' as the target name which is not what we want.

Change-Id: I3f6c536175ae9df1606205fed5f9ee6eb268d631
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240526
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>