[netstack3] Limit size of PMTU cache
Use an LRU cache to limit the number of PMTU entries we maintain. This
prevents a denial-of-service attack where PacketTooBig messages are
repeatedly sent to the stack until it OOMs. We discard the
least-recently-used entry when the cache becomes full; currently this
will just be the entry that was inserted first, since the cache is not
actually *used* for IP fragmentation yet. This property will be more
useful when that is the case (see https://fxbug.dev/404629697) since it
will prioritize in-use paths to remain in the cache.

Fixed: 404634511
Test: netstack3-core-ip-test
Change-Id: I6d01599374a1d1a85513c270812e67b62d4a78e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237566
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>