[fuchsia-controller] Only allow unions with exactly one member
The dynamic Python FIDL bindings for union types enforce only one member
per union, and the variant must be specified upfront (except in the case
of a flexible union with no variant, or when using make_default internally).

The primary motivation for this change was a bug in the implementation
of fuchsia-controller bindings for FIDL methods that take a union for
their arguments. The implementation required positional arguments for
the method, but the union type did not accept keyword arguments.

There is a known bug that still exists even after this change which is
that the comparison of two flexible unions with an unknown variant may
return a nonsense boolean.

Test: Add tests for unions in static and dynamic bindings
Bug: 404285816
Fixed: 42078357
Change-Id: If2a0cf5593e1b8f9007330dd82ad7614c7306f90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1223484
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Sam Balana <sbalana@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>