[fidl][python][fuchsia-controller] Implement FIDL result unwrap for static bindings
Each FIDL result is a union of at most three variants: response,
framework_err, and err. Each variant could be None, and response can
even be None when it's present. The dynamic FIDL bindings have an
unwrap() method to conveniently extract the response field or raise an
exception if it's not present.

This change implements a similar unwrap() method in the static
bindings, and modifies the unwrap() method in the dynamic bindings to
match. This modification is to check framework_err if it's part of the
result type.

There is however still a difference between the two implementations.
Because the static bindings do not compile empty success structs,
unwrap() can return None when the response is an empty success struct.
In the dynamic bindings, unwrap() would return the empty success struct.

Test: Add many tests for unwrapping a FIDL result
Bug: 355717613
Change-Id: I3f238a54af40261ec477b19f59da2ede455deb5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238164
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Marc Khouri <mnck@google.com>