[kernel][vm] Refactor RangeChangeUpdateCowChildren
Makes RangeChangeUpdateCowChildren a static method that takes a
LockedPtr, instead of adopting a Guard. Using a LockedPtr is simpler
for the final implementation and does not impact call sites.

As this method does not have public callers it is moved to the private
section, where the LockedPtr is declared.

Bug: 338300943
Change-Id: I7cbd067dba5151784f137a4c62e32c9781a3d26a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236964
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>