[assembly] Set release version during config generation.
This cleans up the logic that sets the "release_version" field in
board_configs and board_input_bundles. Build rules can set the values
explicitly in GN or Bazel targets, but the config generator tool will
set the default value "unversioned" if no value is provided.

This also allows targets to set these versions as a string directly in
the target, or as a path to a file somewhere in their workspace. This is
necessary to support some board definitions in other repositories.

Bug: b/399808455
Change-Id: I898de09b053fa603f6371afc9e5718058006dc6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238429
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>