[sestarnix] Enable capable_sys test suite test
This requires adding a phoney response to the FIBMAP ioctl. We're
reasonably confident that the SELinux Test Suite is going to be the
only caller ever of the FIBMAP ioctl ever, but we still protect
ourselves by guarding the phoney response behind the selinux_test_suite
kernel option.

Bug: 364569559
Bug: 392615438
Bug: 404795644
Fixes: 398663315
Change-Id: I07e433572e6b803ac60bfb14adb55c6eaff0cc5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240945
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Wez <wez@google.com>