Reland "[aml-thermal-s905d2g-legacy] Use SDK pdev library"
This is a reland of commit 3e0f29b0ce98bb02d172ad553dff557a57cffbaf

The original CL caused the smart display buildbots to fail because the
aml-thermal driver would crash if it failed to get the smc resource.
Before the original CL, the driver would ignore the fact that it could
not get an smc resource.

The original CL has been updated to not crash if it fails to get an smc
resource.

Original change's description:
> [aml-thermal-s905d2g-legacy] Use SDK pdev library
>
> Make the aml-thermal-s905d2g-legacy driver use the
> //sdk/lib/driver/platform-device/cpp SDK library in order to interact
> with the platform device.
>
> Bug: 366517646
> Change-Id: I349d30acff517fe03a1471cf7c2da30c6baefe11
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239304
> Reviewed-by: Andres Oportus <andresoportus@google.com>
> Commit-Queue: Nick Eaton <nickeaton@google.com>

Bug: 366517646
Change-Id: Ic3a20c40a78c8436fe179dc39d8f1710688db464
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241564
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>