[idk][ffx] Add //sdk:host_tools_collection for clidoc
Provide a more intentional target for clidoc to use for identifying
tools for which to create documentation.

This also removes a dependency on a `.sdk` file, which is an internal
implementation detail. clidoc_run.py must still be updated to move
entirely off the internal format.

Bug: 407083737
Change-Id: Iff2150c07163029df23336490abae57704ea0e6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222565
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>