[flash-kola] slight reboot message rewording
"Rebooting to latest bootloader" was causing some confusion as it made
it seem like the bootloader was always out-of-date, even if the dev
hasn't changed anything in their tree.

Removing the "latest" part to make it more clear this is just part of
the normal flashing process and doesn't have anything to do with
out-of-date images.

Bug: b/407614253
Change-Id: Ia656273823768c11e1ed9a078e82a0c24e754c7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240886
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>