[fdf][rust] Fix lock contention on task drop
When dropping the future of a task because the dispatcher is shutting
down, we need to make sure we aren't holding the lock on the currently
active future in case something tries to re-wake the task. Otherwise we
get a lock poisoning crash.

Change-Id: I0ae842d22730f3414607069d70cd4dd82573bb96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241765
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>