[assembly] Set "version" field in board_configurations
fxr/1238429 is adding an optional field to the board_configuration and
board_input_bundle targets named "version", which let you specify the
release version of the given artifact.

Eventually this field will be required.

This CL adds the new field to a handful of tests in the //src/security
directory in the fuchsia tree.

Bug: b/399808455
Change-Id: Ie8c3c17a3d0cf6cc071ce578a84521b97d7976c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240605
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>