[owners] Remove dhobsd@google.com from src/connectivity/network/netlink/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the OWNERs
file tidy.  We ask that you do not ignore this change and approve it
unless you know a reason the OWNER should remain.  It can always be
reverted if needed.

If this change is in error, vote the lowest CR value (i.e. reject the CL)
and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/dhobsd@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iac44e24192bd4a18641db4b81a9e95150d2e9f15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241790
Fuchsia-Auto-Submit: Owners-Cleanup-Prod <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>