[test_pilot] add summaries
This CL adds summary processing to test-pilot. In addition to creating
a default summary for tests that don't produce one, this CL makes
test-pilot merge summaries from multiple sources.

The actual names of directories used in the output directory and the
way we deal with summaries is likely to change as we integrate
test-pilot into the test architecture and finalize the output format.

Test: new unit tests

Bug: 327640651
Change-Id: Ib42273d5617774430f67527655987c1c4936472a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240607
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>