[netstack3] Avoid debug-printing IP addresses
We're currently debug-printing any fuchsia.net.interfaces.admin/Control
request that we process, which, in the case of the AddAddress and
RemoveAddress methods, includes an IP address that will be formatted via
the Debug impl for [u8; 4] or [u8; 16] rather than pretty-printed. This
means it won't be redacted along with other IP addresses. Remove this
debug print since we already log in the processing for most of these
requests anyway.

Change-Id: Ie05a9ec20944e7b378e42b0c4c7b7884a36876d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242086
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>