[sestarnix] Create all Fuchsia-space Anon nodes as private
Migrate all Anon nodes wrapping Fuchsia primitives (those with
internal names of the form "[fuchsia:...]", which do not align
directly with Linux equivalents) to be private, so that they are
un-labeled and not access-checked.

Labeling & checks will be re-introduced for this class of file node
as part of a broader review of Fuchsia-space resource treatment.

Bug: 404739824, 404773987
Change-Id: I66ab5fa07381a8bb9501221982f0f802cace0ab3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242104
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>