[kernel] GenericWatchdot::Pet should check if dog is enabled
Lockup detector calls hw_watchdog_pet that boils down to
GenericWatchdog::Pet in this case.

This lead to some confusing experiments when bringing up the watchdog,
in which the watchdog would be left enabled, but tell the kernel it was
disabled to observe the kernel rebooting after the bite time.

Bug: 407841192
Change-Id: Ic9b249e5d50d2e849ebb11bf61891ff2c67366a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237564
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>