[fxfs] Don't use the root_gen_dir for unicode_gen
This results in rust-analyzer thinking that the set of possible
dependencies for unicode_gen is all of the root_gen_dir, which is
thousands and thousands of files.

Instead, put unicode_gen.rs in its own directory, which lowers the
number of dependent files for the unicode crate from 45324 to just a
handful on a minimal.x64 build.

Test: build still works, existing tests.
Bug: fxbug.dev/270999861
Change-Id: Id1610ed08a87ed9118023987a2a7520e66e26729
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1242964
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: John Wittrock <wittrock@google.com>