[assembly] Allow clobbering dtbo from different assemblies
If the main images and recovery images both add a DTBO via the board
and if recovery uses AB-slotting, then two DTBOs will be destined
for the same A parittion. We allow this as long as the DTBO images
are equivalent.

As DTBOs are small (a few MiB), then hashing the contents is very
fast.

Change-Id: Iebb5a7b76fa461436a0d17e3636c53c2000a1317
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241729
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>