[sestarnix] Account for transitions on fs_use_task
Update `fs_use_task` handling to calculate a new SID using the process
SID as source and target, instead of using the process' SID, to account
for type transitions.

Use the source role as default role when calculating a new SID for
sockets and pipes, as observed in SELinux.

Bug: 364568517, 377912777
Change-Id: I141477d481990dc614855b752b95a99da7dfa725
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225726
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Marina Ciocea <marinaciocea@google.com>