[kernel] Rename ktrace-tests to ktrace-internal-tests
The ktrace tests today actually test the internal KTraceState
implementation. This rename reflects that behavior, and also prepares
us to introduce new tests that test the KTrace singleton's
functionality directly. This will be useful when we start to
implement per-CPU buffers, which will not utilize the internal
KTraceState object.

This is a pure rename and is therefore a functional no-op.

Bug: 404539312
Change-Id: I8cb19a2910759e688d85209c8f048a770120124b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243126
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>