[kernel] Add Reserve method to KTrace
This CL adds a Reserve method to KTrace that wraps the pre-existing
internal::KTraceState::Reserve method. This means KTrace now
implements the fxt::Writer protocol, and the static instance can
now be passed in to the Emit* functions directly.

This is a functional no-op, but will make it easier to add support
for per-CPU buffers in the future.

Bug: 404539312
Change-Id: I6bbee77867382f214a0fe4d2b1105d44913128c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243284
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>