[kernel] Implement SpscBuffer reads and writes
This CL finishes up the implementation of the SpscBuffer by adding
support for reads and writes.

The Read function is relatively straightforward in its behavior, but
notably requires the caller to pass in a CopyOutFn. This argument
allows the usage of different copy functions in different environments.
Concretely, this allows use memcpy in the tests, and will allow us to
use usercopy in the kernel.

The Write operation is a little more complex, since it must be
transactional in nature. To support this, logical Writes are split into
three separate methods:
1. A Reserve method, which verifies that a slot of the requested
   size exists in the ring buffer and returns that slot as a
   Reservation object.
2. A Write method, called on the Reservation object, which actually
   writes data into the pre-reserved slot.
3. A Commit method, called on the Reservation object, which moves the
   write pointer forward and thus makes the write visible to reads.

Bug: 404539312
Change-Id: I00e487afb15051d6d9ddd79cc2612dd9540a801d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235160
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Grossman <johngro@google.com>