[kernel][vm] Unlocked FreePages
Removes the lock requirement on calling FreePages as it is not actually
needed and existed as a 'just in case' for future changes, but none
have yet needed to leverage this being locked. As the lock might be
dropped in between a page being removed from the page_list_ and
actually freed, requiring that the lock be re-acquired just to perform
the free does not provide any value.

Bug: 338300943
Change-Id: I38a58237bfafb3b887845786bafaff819543cc79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239006
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>