[kernel][vm] Add page methods take DeferredOps
Instead of taking a boolean do_range_update changes all the add page
related methods to take a DeferredOps instead. The original meaning of
the boolean is preserved by a nullptr being same as false, and fully
skipping range updates, and a non-null pointer being the same as true
and requiring all range updates (i.e. this and children to be
performed)

Bug: 338300943
Change-Id: I6c505e4358cd2e0dffef08d2d37cab4d81dcfb6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239011
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>