[usb-rs] Manage buffer lifecycles correctly
The previous code unsafely leaked pointers to the reaper thread for the
buffers. Now URBs each have their own associated buffer which is tied to
their lifecycle.

Also we now try to cancel URBs when the waiting future is dropped.

There's still a chance that the reaper thread could successfully reap a
URB but the future could be gone already, in which case the data in the
buffer will be lost. So there's one more potential bug to squash.

Change-Id: I39438522d8d9d2bcd5feebd9b78a727538833ae2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236764
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>