[ffx][repo] Fixing handling no target device
This fixes the early validation that there is a target device
when starting the package server without the `--no-device` option.

Some changes to the structure of the ffx_command_error::Error data
made it impossible to go from a fho::Error to a FfxTargetError in
order to determine the cause.

This change updates the FfxError type to use the underlying FfxTargetError
as the "source" of the error, which allows accessing it and downcast it
correctly.

Bug: 407770719
Change-Id: I2de831df2fd8103f063aed0dc18966514fc565a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243884
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>