[assembly] Add release version numbers to Product configs.
Product configs already include version information in the form
of a file.

For consistency, and to remove the need to read that file during
bisection, we copy the contents of that file to the top level of the
"product" section of the config json.

Bug: b/399808455
Change-Id: I7ec08187f925162cd7dc6c9c4c137564f1649ece
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238666
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>