[starnix] Have WRITE permission for set_content_size
Modifying the content size is a modification to the VMO and is supposed
to require the WRITE permission, as per RFC-0238. This changes the
SLICE, which is not writable, to a writable clone, which will allow the
kernel to begin enforcing the permission.

Change-Id: Ibbe1e7291141303684484b098cbc5d40a70ed080
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241186
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Perry <adamperry@google.com>