[starnix][zxio] Add function to update atime when closing
We now have a way to inform the underlying filesystem that a node is
pending an atime update (due to file access) by sending a fuchsia.io
Node Get_Attributes request with `PENDING_ACCESS_TIME_UPDATE` query.
When we close a node, we may want to update atime first before closing
it.

This will be used in follow-up CL to have persistent update of atime in
Starnix.

Bug: 401073280
Change-Id: Iccd6375db053358bcefb187f84e8b65da01ed03e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235947
Commit-Queue: Rui Qi Sim <rqsim@google.com>
Reviewed-by: James Robinson <jamesr@google.com>