[starnix] atime updated persistently
When a node is accessed, it is marked that it is pending a time_access
update. When node info is requested for remote filesystems, Starnix will
add on an additional node attributes query, `PENDING_ACCESS_TIME_UPDATE`,
if there is a pending time access update to the get_attributes request
sent to the underlying filesystem. This should inform the filesystem
that a file has been accessed and to update access time before returning
the node attributes.

This change impacts only accesses within Starnix.

This fixes atime related gvisor tests. The tests were ran with caching
turned off to test for the persistent updates.

Bug: 401073280
Test: fx test starnix_gvisor_stat_times_test_on_fxfs
Test: fx test starnix_gvisor_utimes_test_on_fxfs
Test: fx test starnix_gvisor_symlink_test_on_fxfs
Change-Id: I958460a5f6969c46a3f0c4836e15ac93a275b0b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203887
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>