[realm_builder] fix race in realm builder event stream test
The realm builder test for event streams started two eager components,
with one connecting to an event stream and expecting to see a start
event for the other. As these components start up at the same point,
sometimes the former would start before the latter could connect to its
event stream. In this case, we fully expect the second component to miss
the event, as it happened before the event stream was opened.

This commit fixes this race by making the first component lazy and
having it be started by the second component only once the event stream
is connected to.

Change-Id: I5ff89328029ada7f08cb42e760f542030369d802
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1243844
Reviewed-by: Olivia Nordquist <nolivia@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>