[netstack3] Prepare congestion module for SACK
Rework the congestion module to operate based on the SACK scoreboard to
estimate how many bytes are in flight, which falls back to the same
estimation when there are no SACK blocks.

We delegate the decision of which segment to send to the congestion
module, along with calculating the congestion limit and mss.

While I'm here, rework a bit the counter semantics so they better
reflect the retransmission cases.

Introduce a test that verifies that congestion control is in fact
limiting the amount of data sent per RTT by evaluating the average
flight size when not in slow start.

Bug: 42078221
Change-Id: I4019dcbce85e749017e5219892b32a18649f1937
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1239886
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>