[sdk][fidl][gpio] Define WAKEABLE interrupt option.
Previously we passed through the ZX_INTERRUPT_WAKE_VECTOR kernel flag
to syscalls, but this didn't work for all pinimpl drivers. For
example, some pinimpl drivers demux the wakeable interrupt to virtual
interrupts, where the kernel flag should not be passed through.

This change defines a bit "WAKEABLE" which is not tied to the kernel
flag. An interrupt is "WAKEABLE" if it can wakes up the system if the
system is suspended.

Bug: 361851116
Bug: 380120239
Change-Id: I48ce3dd13840f83fc9f2029590a978d8cd3dde48
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1240887
API-Review: Christopher Anderson <cja@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>