[sestarnix] Rename the object class and permission types
Kernel-specific class enum is now KernelClass, and the kernel-
specific permission enum KernelPermission.

The kernel-or-other enum is renamed to simple ObjectClass, to
reflect that it refers to a policy class that may or may not be
one of which the kernel is aware.

Change-Id: I92c8af46966288ef1877d0821923aeed810bb008
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1245284
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Wez <wez@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>