[ffx][usb] Add echo packets
This adds echo packets to the usb vsock implementation library and
automatically replies to them. This will be used for probing the usb
connection at synchronization, but might be used for other checks to
validate that the usb connection is still in a functional state.

Bug: 406262417
Change-Id: Ica983cecb38e4ac0bef633c8772c1cc0d2e7d87d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1245327
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>