[fdomain] Try to handle datagram sockets correctly
`zx_socket_read` specifies that when a socket is in datagram mode, a
read returns as many bytes as the buffer allows and then discards the
rest of the datagram.

This became a problem when we started allowing the client to drop a read
future, as there's no longer a 1:1 correlation between read calls on the
host and read syscalls on the target. This means the length specified
when you do a read may not be the length given to the read call.

We mitigate this by simply endeavoring to always return a full datagram
from the FDomain, ignoring the read argument, and then implementing the
truncation on the client side when we're past the IO phase and know the
users intentions.

The one gap in this strategy is that the FDomain can't know the socket
is a datagram socket if we don't have the `INSPECT` right. In such a
case things will probably still work if the client doesn't drop read
futures or uses a consistent datagram size. Dropping read futures is
pretty rare behavior anyway, so the bug should be hard to trigger.

Fixed: b/388039830
Change-Id: Ief3342499d65c06c35941b2d3eb2ed5f18de4cb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217686
Commit-Queue: Casey Dahlin <sadmac@google.com>
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>