[audio][test] Distinct audio driver test suites
The audio_driver_tests suite runs non-hermetically, testing
the system instances of all audio drivers detected.

This CL splits audio_driver_tests into three individually-
invocable packages:
* audio_driver_basic_tests,
* audio_driver_admin_tests, and
* audio_driver_realtime_tests.

The first could be run on any CQ buildbot. It is intended to
be run even when audio_core or audio_device_registry is
already running and connected to the audio drivers found in
devfs. These cases can be added to either of the other two
suites, by appending the flag '--basic'.

The second is intended to be run only on buildbots without
audio_core or audio_device_registry, as it needs exclusive
access to the audio driver. It can run on emulators or (non-
hermetically) real boards such as NUC, VIM3, etc). These
cases can be added to either of the other two suites, by
appending the flag '--admin'.

The third is also intended to be run only in the absence of
audio services (audio_core or audio_device_registry).
Additionally, it should only be run on real boards (NUC,
VIM3, etc.) because it must be guaranteed to execute in a
real-time environment, on audio drivers that are operating
real audio hardware. These cases can be added to either of
the other two suites, by appending the flag '--position'.
Note that many drivers (including virtual-audio) do not yet
pass all these cases, so they will not be enabled by default
until those drivers are fixed.

Bug: 370741565

Change-Id: I8092222356f5a5c43d6bd3e04e7b41594b1bdc67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1241090
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>