[ffx] Remove hardcoded sdk/manifest/core
This removes the reliance on the assumed existence of sdk/manifes/core in-tree.

The IDK/SDK building is being improved, and the tools used within Fuchsia need to be
flexible enough to allow the IDK to be moved, and most tool executions not require the
IDK manifest to find other ffx- subtools, and the related host-tools.

Bug: 397989792

Change-Id: I92c295cc99aea006629f9bf48482e24479ce1ceb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221424
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>