[starnix] Always decrement counter when timer is cancelled
Previously, when we were using an event, the timer that was cancelling
the outstanding timer would clear the event signal for both the new and
the old timer.

Since we are now using a counter, each message has to be acknowledged
with an explicit decrement.

Change-Id: Ibace4418a8e15972f04e82a139cfd6a693b4e51d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1245345
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>