[build][api] Fix --last-build-only implementation.
Fix the logic used to implement `--last-build-only` by:

- Using the new Ninja `outputs` tool to get the set of artifacts
  from the last build invocation.

- Use the `inputs` tool to get the set of sources used by the
  last build invocation.

- Use both sets to filter build API modules. This ensures that
  both prebuilts (sources) and intermediate artifacts (e.g.
  unstripped binaries) are correctly listed.

NOTE: The bug is not fixed because prebuilt binaries, like Clang
and Rust runtime files are never part of the result. The reason is
that they are not listed as proper implicit inputs by Ninja targets.
This will be solved in a separate CL that adds the proper data_deps
to C++ and Rust binary targets.

Bug: 408504040
Change-Id: I444a980ebdf439f7069d28d869a23603a9004be4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1246784
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>